+++ VisaCheckout/src/main/java/com/braintreepayments/api/VisaCheckoutCreateProfileBuilderCallback.java
@@ -1,0 +1,7 @@
+package com.braintreepayments.api;
+
+import com.visa.checkout.Profile;
+
+public interface VisaCheckoutCreateProfileBuilderCallback {
+    void onResult(Profile.ProfileBuilder profileBuilder, Exception error);
+}