+++ VisaCheckout/src/main/java/com/braintreepayments/api/VisaCheckoutOnActivityResultCallback.java
@@ -1,0 +1,5 @@
+package com.braintreepayments.api;
+
+public interface VisaCheckoutOnActivityResultCallback {
+    void onResult(PaymentMethodNonce paymentMethodNonce, Exception error);
+}