+++ VisaCheckout/src/main/java/com/braintreepayments/api/VisaCheckoutTokenizeCallback.java
@@ -1,0 +1,5 @@
+package com.braintreepayments.api;
+
+public interface VisaCheckoutTokenizeCallback {
+    void onResult(PaymentMethodNonce paymentMethodNonce, Exception error);
+}