+++ VisaCheckout/src/test/java/com/braintreepayments/api/TestActivity.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api.test;
+package com.braintreepayments.api;
 
 import android.os.Bundle;