+++ VisaCheckout/src/test/java/com/braintreepayments/api/VisaCheckoutBuilderUnitTest.java
@@ -1,21 +1,39 @@
-package com.braintreepayments.api.models;
-
-import android.os.Parcel;
+package com.braintreepayments.api;
 
 import com.visa.checkout.VisaPaymentSummary;
 
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PowerMockIgnore;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.rule.PowerMockRule;
 import org.robolectric.RobolectricTestRunner;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.skyscreamer.jsonassert.JSONCompareMode;
 
 import static junit.framework.Assert.assertEquals;
+import static org.powermock.api.mockito.PowerMockito.when;
 
 @RunWith(RobolectricTestRunner.class)
+@PowerMockIgnore({"org.powermock.*", "org.mockito.*", "org.robolectric.*", "android.*", "androidx.*"})
+@PrepareForTest({ VisaPaymentSummary.class })
 public class VisaCheckoutBuilderUnitTest {
+
+    @Rule
+    public PowerMockRule mPowerMockRule = new PowerMockRule();
+
+    private VisaPaymentSummary visaPaymentSummary;
+
+    @Before
+    public void beforeEach() throws Exception {
+        visaPaymentSummary = PowerMockito.mock(VisaPaymentSummary.class);
+        PowerMockito.whenNew(VisaPaymentSummary.class).withAnyArguments().thenReturn(visaPaymentSummary);
+    }
 
     @Test
     public void build_withNullVisaPaymentSummary_buildsEmptyPaymentMethod() throws JSONException {
@@ -31,20 +49,15 @@ public class VisaCheckoutBuilderUnitTest {
 
     @Test
     public void build_withVisaPaymentSummary_buildsExpectedPaymentMethod() throws JSONException {
+        when(visaPaymentSummary.getCallId()).thenReturn("stubbedCallId");
+        when(visaPaymentSummary.getEncKey()).thenReturn("stubbedEncKey");
+        when(visaPaymentSummary.getEncPaymentData()).thenReturn("stubbedEncPaymentData");
+
         JSONObject base = new JSONObject();
         JSONObject paymentMethodNonceJson = new JSONObject();
 
-        JSONObject summaryJson = new JSONObject()
-                .put("encPaymentData", "stubbedEncPaymentData")
-                .put("encKey", "stubbedEncKey")
-                .put("callid", "stubbedCallId");
-
-        Parcel in = Parcel.obtain();
-        in.writeString("SUCCESS");
-        in.writeString(summaryJson.toString());
-        in.setDataPosition(0);
-
-        VisaPaymentSummary visaPaymentSummary = VisaPaymentSummary.CREATOR.createFromParcel(in);
+        VisaCheckoutBuilder visaCheckoutBuilder = new VisaCheckoutBuilder(visaPaymentSummary);
+        visaCheckoutBuilder.build(base, paymentMethodNonceJson);
 
         JSONObject expectedBase = new JSONObject();
         JSONObject expectedPaymentMethodNonce = new JSONObject();
@@ -52,9 +65,6 @@ public class VisaCheckoutBuilderUnitTest {
         expectedPaymentMethodNonce.put("encryptedKey", "stubbedEncKey");
         expectedPaymentMethodNonce.put("encryptedPaymentData", "stubbedEncPaymentData");
         expectedBase.put("visaCheckoutCard", expectedPaymentMethodNonce);
-
-        VisaCheckoutBuilder visaCheckoutBuilder = new VisaCheckoutBuilder(visaPaymentSummary);
-        visaCheckoutBuilder.build(base, paymentMethodNonceJson);
 
         JSONAssert.assertEquals(expectedBase, base, JSONCompareMode.STRICT);
     }